DKC Studios Section is ready ADDED: DKC Studios Directory DKC Studios/index.html
CHANGED:
uniform.css now has bSection which is section unstylized.
bSection creates a grid layout to organize the display of games on DKC Studios/index.html

NOTES:
Files are not created for:
Project Valkyrie
Project Blue
Project Green

These need to be implemented and connected to the a href in the DKC Studios #bSection section a.